polish some components
