+++ apps/website/src/pages/index.astro
@@ -281,12 +281,9 @@ import Layout from "../layouts/Layout.astro";
         >
           <h2 class="text-[3.75rem] leading-none text-text">About</h2>
           <p class="text-body leading-[220%] text-text">
-            The Bible industry is full of copyright, paid software, non-private,
-            or non-modern software. We want to fix that, and comply with Jesus'
-            command in Matthew 10:8, "Freely you have received; freely give." In
-            the future we also want to help churches and communities with our
-            free software.
+             Bibleio is a Christian organization based on making innovative, modern, free and open-source Bible technology. The Bible industry is full of copyright, paid resources, & clunky old software. We want to fix that. 
           </p>
+          <a href="/readme"><Button>View README</Button></a>
         </div>
 
         <div
@@ -299,7 +296,7 @@ import Layout from "../layouts/Layout.astro";
             Take a look at our free & open-source Bible API, our new formatting
             standard, BibleJSON, and our beautiful design system.
           </p>
-          <a href="https://dev.bibleio.com"><Button>Look</Button></a>
+          <Button disabled>Coming soon</Button>
         </div>
         <div
           class="flex h-[24rem] flex-col gap-24 border-stroke p-48 max-[580px]:border-t max-[580px]:p-24 max-[580px]:text-h2"
@@ -311,8 +308,8 @@ import Layout from "../layouts/Layout.astro";
             Come talk with us, contribute, and give feedback!
           </p>
           <div class="flex gap-12">
-            <a href="https://github.com/bibleio"><Button>GitHub</Button></a>
             <a href="https://discord.gg/7eVCyQ5GGb"><Button>Discord</Button></a>
+            <a href="https://github.com/bibleio"><Button>GitHub</Button></a>
             <a href="/contact"><Button variant="secondary">Contact</Button></a>
           </div>
         </div>
