+++ packages/design/src/components/base/Badge/Badge.stories.tsx
@@ -4,31 +4,31 @@ import React from "react";
 import { Badge, type BadgeProps } from "./Badge";
 
 const meta: Meta<typeof Badge> = {
-	component: Badge,
-	title: "Badge",
-	tags: ["autodocs"],
-	argTypes: {
-		variant: {
-			options: [
-				"primary",
-				"secondary",
-				"success",
-				"warning",
-				"danger",
-				"blue",
-				"purple",
-			],
-			control: { type: "radio" },
-		},
-	},
+  component: Badge,
+  title: "Base/Badge",
+  tags: ["autodocs"],
+  argTypes: {
+    variant: {
+      options: [
+        "primary",
+        "secondary",
+        "success",
+        "warning",
+        "danger",
+        "blue",
+        "purple",
+      ],
+      control: { type: "radio" },
+    },
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof Badge>;
 
 export const Default: Story = {
-	render: (args: BadgeProps) => <Badge {...args}>Badge</Badge>,
+  render: (args: BadgeProps) => <Badge {...args}>Badge</Badge>,
 };
 Default.args = {
-	variant: "primary",
+  variant: "primary",
 };
