+++ packages/design/src/components/base/Checkbox/Checkbox.stories.tsx
@@ -3,17 +3,17 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { Checkbox } from "./Checkbox";
 
 const meta: Meta<typeof Checkbox> = {
-	component: Checkbox,
-	title: "Checkbox",
-	tags: ["autodocs"],
+  component: Checkbox,
+  title: "Base/Checkbox",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof Checkbox>;
 
 export const Default: Story = {
-	render: (args) => <Checkbox {...args} />,
+  render: (args) => <Checkbox {...args} />,
 };
 Default.args = {
-	disabled: false,
+  disabled: false,
 };
