+++ packages/design/src/components/base/InputOTP/InputOTP.stories.tsx
@@ -3,15 +3,15 @@ import React from "react";
 import { InputOTP, type InputOTPProps } from "./InputOTP";
 
 const meta: Meta<typeof InputOTP> = {
-	component: InputOTP,
-	title: "Input OTP",
-	tags: ["autodocs"],
+  component: InputOTP,
+  title: "Base/Input OTP",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof InputOTP>;
 
 export const Default: Story = {
-	render: (args: InputOTPProps) => <InputOTP {...args} />,
+  render: (args: InputOTPProps) => <InputOTP {...args} />,
 };
 Default.args = { disabled: false, maxLength: 6 };
