+++ packages/design/src/components/base/Loader/Loader.stories.tsx
@@ -3,14 +3,14 @@ import React from "react";
 import { Loader } from "./Loader";
 
 const meta: Meta<typeof Loader> = {
-	component: Loader,
-	title: "Loader",
-	tags: ["autodocs"],
+  component: Loader,
+  title: "Base/Loader",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof Loader>;
 
 export const Default: Story = {
-	render: () => <Loader />,
+  render: () => <Loader />,
 };
