+++ packages/design/src/components/base/Note/Note.stories.tsx
@@ -3,21 +3,21 @@ import React from "react";
 import { DangerNote, InfoNote, SuccessNote, WarningNote } from "./Note";
 
 const meta: Meta<typeof InfoNote> = {
-	component: InfoNote,
-	title: "Notes",
-	tags: ["autodocs"],
+  component: InfoNote,
+  title: "Base/Notes",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof InfoNote>;
 
 export const Default: Story = {
-	render: () => (
-		<div className="flex flex-col gap-24">
-			<InfoNote>Note</InfoNote>
-			<SuccessNote>Note</SuccessNote>
-			<WarningNote>Note</WarningNote>
-			<DangerNote>Note</DangerNote>
-		</div>
-	),
+  render: () => (
+    <div className="flex flex-col gap-24">
+      <InfoNote>Note</InfoNote>
+      <SuccessNote>Note</SuccessNote>
+      <WarningNote>Note</WarningNote>
+      <DangerNote>Note</DangerNote>
+    </div>
+  ),
 };
