+++ packages/design/src/components/base/Select/Select.stories.tsx
@@ -3,56 +3,56 @@ import React from "react";
 
 import type { SelectProps } from "../Select";
 import {
-	Select,
-	SelectGroup,
-	SelectItem,
-	SelectLabel,
-	SelectSeparator,
+  Select,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectSeparator,
 } from "./Select";
 
 const meta: Meta<typeof Select> = {
-	component: Select,
-	title: "Select",
-	tags: ["autodocs"],
+  component: Select,
+  title: "Base/Select",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof Select>;
 
 export const Default: Story = {
-	render: (args: SelectProps) => (
-		<Select {...args}>
-			<SelectGroup>
-				<SelectLabel>Fruits</SelectLabel>
-				<SelectItem value="apple">Apple</SelectItem>
-				<SelectItem value="banana">Banana</SelectItem>
-				<SelectItem value="bluebeSelectPropsrry">Blueberry</SelectItem>
-				<SelectItem value="grapes">Grapes</SelectItem>
-				<SelectItem value="pineapple">Pineapple</SelectItem>
-			</SelectGroup>
-			<SelectSeparator />
-			<SelectGroup>
-				<SelectLabel>Vegetables</SelectLabel>
-				<SelectItem value="eggplant">Eggplant</SelectItem>
-				<SelectItem value="broccoli">Broccoli</SelectItem>
-				<SelectItem value="carrot" disabled>
-					Carrot
-				</SelectItem>
-				<SelectItem value="zucchini">Zucchini</SelectItem>
-				<SelectItem value="leek">Leek</SelectItem>
-			</SelectGroup>
-			<SelectSeparator />
-			<SelectGroup>
-				<SelectLabel>Meat</SelectLabel>
-				<SelectItem value="beef">Beef</SelectItem>
-				<SelectItem value="chicken">Chicken</SelectItem>
-				<SelectItem value="lamb">Lamb</SelectItem>
-				<SelectItem value="pork">Pork</SelectItem>
-			</SelectGroup>
-		</Select>
-	),
+  render: (args: SelectProps) => (
+    <Select {...args}>
+      <SelectGroup>
+        <SelectLabel>Fruits</SelectLabel>
+        <SelectItem value="apple">Apple</SelectItem>
+        <SelectItem value="banana">Banana</SelectItem>
+        <SelectItem value="bluebeSelectPropsrry">Blueberry</SelectItem>
+        <SelectItem value="grapes">Grapes</SelectItem>
+        <SelectItem value="pineapple">Pineapple</SelectItem>
+      </SelectGroup>
+      <SelectSeparator />
+      <SelectGroup>
+        <SelectLabel>Vegetables</SelectLabel>
+        <SelectItem value="eggplant">Eggplant</SelectItem>
+        <SelectItem value="broccoli">Broccoli</SelectItem>
+        <SelectItem value="carrot" disabled>
+          Carrot
+        </SelectItem>
+        <SelectItem value="zucchini">Zucchini</SelectItem>
+        <SelectItem value="leek">Leek</SelectItem>
+      </SelectGroup>
+      <SelectSeparator />
+      <SelectGroup>
+        <SelectLabel>Meat</SelectLabel>
+        <SelectItem value="beef">Beef</SelectItem>
+        <SelectItem value="chicken">Chicken</SelectItem>
+        <SelectItem value="lamb">Lamb</SelectItem>
+        <SelectItem value="pork">Pork</SelectItem>
+      </SelectGroup>
+    </Select>
+  ),
 };
 Default.args = {
-	label: "Select a food",
-	disabled: false,
+  label: "Select a food",
+  disabled: false,
 };
