+++ packages/design/src/components/base/Switch/Switch.stories.tsx
@@ -3,17 +3,17 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { Switch } from "./Switch";
 
 const meta: Meta<typeof Switch> = {
-	component: Switch,
-	title: "Switch",
-	tags: ["autodocs"],
+  component: Switch,
+  title: "Base/Switch",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof Switch>;
 
 export const Default: Story = {
-	render: (args) => <Switch {...args} />,
+  render: (args) => <Switch {...args} />,
 };
 Default.args = {
-	disabled: false,
+  disabled: false,
 };
