+++ packages/design/src/components/base/Tooltip/Tooltip.stories.tsx
@@ -3,34 +3,34 @@ import React from "react";
 
 import { Button } from "../Button";
 import {
-	Tooltip,
-	TooltipContent,
-	type TooltipProps,
-	TooltipProvider,
-	TooltipTrigger,
+  Tooltip,
+  TooltipContent,
+  type TooltipProps,
+  TooltipProvider,
+  TooltipTrigger,
 } from "./Tooltip";
 
 const meta: Meta<typeof Tooltip> = {
-	component: Tooltip,
-	title: "Tooltip",
-	tags: ["autodocs"],
+  component: Tooltip,
+  title: "Base/Tooltip",
+  tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj<typeof Tooltip>;
 
 export const Default: Story = {
-	render: (args: TooltipProps) => (
-		<TooltipProvider>
-			<Tooltip {...args}>
-				<TooltipTrigger>
-					<Button>Hover me</Button>
-				</TooltipTrigger>
-				<TooltipContent>Tooltip</TooltipContent>
-			</Tooltip>
-		</TooltipProvider>
-	),
+  render: (args: TooltipProps) => (
+    <TooltipProvider>
+      <Tooltip {...args}>
+        <TooltipTrigger>
+          <Button>Hover me</Button>
+        </TooltipTrigger>
+        <TooltipContent>Tooltip</TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
+  ),
 };
 Default.args = {
-	delayDuration: 0,
+  delayDuration: 0,
 };
