+++ packages/design/src/components/index.ts
@@ -6,7 +6,6 @@ export * from "./base/RadioGroup";
 export * from "./base/Select";
 export * from "./base/Slider";
 export * from "./base/Switch";
-export * from "./base/Brand";
 export * from "./base/ListItem";
 export * from "./base/Progress";
 export * from "./base/Loader";
@@ -14,10 +13,13 @@ export * from "./base/Tooltip";
 export * from "./base/InputOTP";
 export * from "./base/Badge";
 export * from "./base/Tabs";
-export * from "./base/Shapes";
 export * from "./base/Dialog";
 export * from "./base/AlertDialog";
 export * from "./base/Note";
+
+// Brand stuff
+export * from "./brand/Shapes";
+export * from "./brand/Brand";
 
 // Reused website components
 export { default as Footer } from "./website/Footer.astro";
