+++ packages/design/src/components/website/Footer.astro
@@ -1,5 +1,6 @@
 ---
-import { Logo } from "../base/Brand";
+
+import { Logo, IconLogo } from "../brand/Brand";
 import ThemeSwitcher from "./ThemeSwitcher.astro";
 
 /**
@@ -22,18 +23,7 @@ const { hrefTargetOutside, devHrefTargetOutside } = Astro.props;
     class="absolute left-1/2 top-[-6%] flex size-48 -translate-x-1/2 items-center justify-center rounded-full border border-stroke bg-bg-1 duration-150 ease-out hover:scale-105 max-[1100px]:top-[-2.2%]"
     href={hrefTargetOutside ? "https://bibleio.com/" : "/"}
   >
-    <svg
-      width="1.375rem"
-      height="2rem"
-      viewBox="0 0 22 32"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      class="fill-accent-reversed"
-    >
-      <path
-        d="M8.09789 1.38197C8 1.68323 8 2.05549 8 2.8V8L2.8 8C2.05549 8 1.68323 8 1.38197 8.09789C0.77309 8.29572 0.295723 8.77309 0.0978869 9.38196C0 9.68323 0 10.0555 0 10.8V11.2C0 11.9445 0 12.3168 0.0978868 12.618C0.295723 13.2269 0.77309 13.7043 1.38197 13.9021C1.68323 14 2.05549 14 2.8 14H8V29.2C8 29.9445 8 30.3168 8.09789 30.618C8.29572 31.2269 8.77309 31.7043 9.38197 31.9021C9.68323 32 10.0555 32 10.8 32H11.2C11.9445 32 12.3168 32 12.618 31.9021C13.2269 31.7043 13.7043 31.2269 13.9021 30.618C14 30.3168 14 29.9445 14 29.2V14H19.2C19.9445 14 20.3168 14 20.618 13.9021C21.2269 13.7043 21.7043 13.2269 21.9021 12.618C22 12.3168 22 11.9445 22 11.2V10.8C22 10.0555 22 9.68323 21.9021 9.38197C21.7043 8.77309 21.2269 8.29572 20.618 8.09789C20.3168 8 19.9445 8 19.2 8L14 8V2.8C14 2.05549 14 1.68323 13.9021 1.38197C13.7043 0.77309 13.2269 0.295723 12.618 0.097887C12.3168 0 11.9445 0 11.2 0H10.8C10.0555 0 9.68323 0 9.38197 0.097887C8.77309 0.295723 8.29572 0.77309 8.09789 1.38197Z"
-      ></path>
-    </svg>
+    <IconLogo size="2rem" reverseColor />
   </a>
   <div class="flex h-full flex-col gap-16 border-l border-stroke p-32">
     <div class="flex items-center justify-between">
